move svgs to public
