+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
+import reactLogo from '../public/react.svg'
 import './App.css'
 import { BrowserRouter, Route,Routes } from 'react-router-dom'
 import NavBar from './components/NavBar'
