+++ src/components/Careers/CareerBox.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useState } from 'react';
 import Box from './textBoxes';
+import ArrowIcon from '../../../public/Arrow.svg'
 
 const CareerBox = (props) => {
   const { title, description } = props;
@@ -26,7 +27,7 @@ const CareerBox = (props) => {
           <Box></Box>
           <div style={{ position: 'absolute', bottom: '20px', right: '30px'}}> {/* Add absolute positioning to child div */}
           <button className="text-white border border-white rounded-full flex items-center w-32 h-10">
-      <img src="src/assets/Arrow.svg" alt="Arrow" class="h-6 w-6 ml-5" />
+          <img src={ArrowIcon} alt="Arrow" class="h-6 w-6 ml-5" />
       <span className = "flex-grow text-right mr-6">Apply</span>
               </button>
               </div>
@@ -39,7 +40,7 @@ const CareerBox = (props) => {
   <Box></Box>
   <div style={{ position: "absolute", bottom: "20px", right: "30px" }}>
     <button className="text-white border border-white rounded-full flex items-center w-32 h-10">
-      <img src="src/assets/Arrow.svg" alt="Arrow" class="h-6 w-6 ml-5" />
+      <img src={ArrowIcon} alt="Arrow" class="h-6 w-6 ml-5" />
       <span className = "flex-grow text-right mr-6">Apply</span>
     </button>
   </div>
