+++ src/components/Home/Home.jsx
@@ -8,7 +8,7 @@ export default function Home(props) {
         <div className="overflow-y-auto flex-auto overflow-hidden w-full h-screen bg-blue-bg">
             <div className="flex">
                 <CallToAction/>
-                <img className="object-scale-down w-[0rem] md:w-[18rem] lg:w-[18rem]  ml-auto mr-24 mt-14" src = "src/assets/HomePic.svg" alt="Invictuss logo"/>
+                <img className="object-scale-down w-[0rem] md:w-[18rem] lg:w-[18rem]  ml-auto mr-24 mt-14" src = "../../../public/HomePic.svg" alt="Invictuss logo"/>
             </div>
             
             <HomeValues/>
