+++ src/components/Home/HomeValues.jsx
@@ -10,21 +10,21 @@ export default function HomeValues(props) {
       <div className="lg:flex justify-center">
       <div className="pt-6 pl-10 pr-10 lg:gap-x-14 ">
           <ValueBox
-            imageSrc="src/assets/Profile.svg"
+            imageSrc="../../../public/Profile.svg"
             title="Security"
             description=" We prioritize the protection of our clients' data through cutting-edge security practices and constant monitoring for emerging threats."
           />
         </div>
         <div className="pt-6 pl-10 pr-10">
           <ValueBox
-            imageSrc="src/assets/UserProfile.svg"
+            imageSrc="../../../public/UserProfile.svg"
             title="Expertise"
             description="We are cybersecurity experts who stay up-to-date with the latest trends and technologies to provide optimal solutions."
           />
         </div>
         <div className="pt-6 pl-10 pr-10 ">
           <ValueBox
-            imageSrc="src/assets/CloudSecurity.svg"
+            imageSrc="../../../public/CloudSecurity.svg"
             title="Trust"
             description="We prioritize building strong relationships with our clients based on transparency, honesty, and ethical behavior."
           />
